requirements.txt
streamlit==1.38.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.3.1
numpy==1.26.4
scikit-learn==1.5.1
rapidfuzz==3.9.6
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
ragas==0.1.14
matplotlib==3.9.0
transformers==4.44.2
torch>=2.2.0
pdfplumber==0.11.4
pandas==2.2.2
regex==2024.5.15